mediarotator/requirements.txt
requests
python-dotenv
PyYAML
textual
fastapi
uvicorn
